Task_Control update

+++ Service/imu_data.h
@@ -40,8 +40,8 @@ typedef struct
     int16_t gyr_int_lst[Axis_Sum];
     int16_t acc_int_lst[Axis_Sum];
 
-    uint8_t gyr_blunt_cnt[Axis_Sum];
-    uint8_t acc_blunt_cnt[Axis_Sum];
+    uint16_t gyr_blunt_cnt[Axis_Sum];
+    uint16_t acc_blunt_cnt[Axis_Sum];
 } IMUData_TypeDef;
 
 typedef struct
